productAPI/requirements.txt
asgiref==3.5.2
cffi==1.15.1
cryptography==38.0.1
Django==4.1
django-urlcrypt==0.1.6
djangorestframework==3.14.0
Pillow==9.2.0
pycparser==2.21
PyMySQL==1.0.2
pytz==2022.2.1
sqlparse==0.4.2
tzdata==2022.2